show warning dialogs when closing unsaved tabs

+++ apps/studio/src/components/CoreTabs.vue
@@ -207,6 +207,7 @@ import { DatabaseEntity } from "@/lib/db/models"
 import PendingChangesButton from './common/PendingChangesButton.vue'
 
 import { safeSqlFormat as safeFormat } from '@/common/utils';
+import pluralize from 'pluralize'
 
 export default Vue.extend({
   props: ['connection'],
@@ -253,7 +254,7 @@ export default Vue.extend({
     }
   },
   computed: {
-    ...mapState('tabs', { 'activeTab': 'active' }),
+    ...mapState('tabs', { 'activeTab': 'active', 'tabs': 'tabs' }),
     ...mapGetters({ 'menuStyle': 'settings/menuStyle', 'dialect': 'dialect' }),
     tabIcon() {
       return {
@@ -297,15 +298,6 @@ export default Vue.extend({
         { event: AppEvent.deleteDatabaseElement, handler: this.deleteDatabaseElement },
         { event: AppEvent.dropDatabaseElement, handler: this.dropDatabaseElement },
         { event: AppEvent.duplicateDatabaseTable, handler: this.duplicateDatabaseTable },
-      ]
-    },
-    contextOptions() {
-      return [
-        { name: "Close", slug: 'close', handler: ({ item }) => this.close(item) },
-        { name: "Close Others", slug: 'close-others', handler: ({ item }) => this.closeOther(item) },
-        { name: 'Close All', slug: 'close-all', handler: this.closeAll },
-        { name: "Close Tabs to Right", slug: 'close-to-right', handler: ({ item }) => this.closeToRight(item) },
-        { name: "Duplicate", slug: 'duplicate', handler: ({ item }) => this.duplicate(item) }
       ]
     },
     lastTab() {
@@ -471,20 +463,6 @@ export default Vue.extend({
     openContextMenu(event, item) {
       this.contextEvent = { event, item }
     },
-    contextClick({ option, item }) {
-      switch (option.slug) {
-        case 'close':
-          return this.close(item)
-        case 'close-others':
-          return this.closeOther(item)
-        case 'close-all':
-          return this.closeAll();
-        case 'close-to-right':
-          return this.closeToRight(item);
-        case 'duplicate':
-          return this.duplicate(item);
-      }
-    },
     async setActiveTab(tab) {
       await this.$store.dispatch('tabs/setActive', tab)
     },
@@ -635,7 +613,15 @@ export default Vue.extend({
         }
       }
     },
-    async close(tab) {
+    async close(tab: OpenTab) {
+      if (tab.unsavedchanges) {
+        const confirmed = await this.$confirm(
+          `really close [icon here] ${tab.title}?`,
+          "you will lose unsaved changes"
+        )
+        if (!confirmed) return
+      }
+
       if (this.activeTab === tab) {
         if (tab === this.lastTab) {
           this.previousTab()
@@ -648,22 +634,47 @@ export default Vue.extend({
         await this.$store.dispatch('data/queries/reload', tab.queryId)
       }
     },
-    closeAll() {
+    async closeAll() {
+      const unsavedTabs = this.tabs.filter((tab) => tab.unsavedChanges)
+      if (unsavedTabs.length > 0) {
+        const confirmed = await this.$confirm(
+          'Close all tabs?',
+          `You have ${unsavedTabs.length} unsaved ${pluralize('tab', unsavedTabs.length)}. Are you sure?`
+        )
+        if (!confirmed) return
+      }
       this.$store.dispatch('tabs/unload')
     },
-    closeOther(tab) {
+    async closeOther(tab: OpenTab) {
       const others = _.without(this.tabItems, tab)
+      const unsavedTabs = others.filter((t) => t.unsavedChanges)
+      if (unsavedTabs.length > 0) {
+        const confirmed = await this.$confirm(
+          'Close other tabs?',
+          `You have ${unsavedTabs.length} unsaved ${pluralize('tab', unsavedTabs.length)}. Are you sure?`
+        )
+        if (!confirmed) return
+      }
+
       this.$store.dispatch('tabs/remove', others)
       this.setActiveTab(tab)
       if (tab.queryId) {
         this.$store.dispatch('data/queries/reload', tab.queryId)
       }
     },
-    closeToRight(tab) {
+    async closeToRight(tab: OpenTab) {
       const tabIndex = _.indexOf(this.tabItems, tab)
       const activeTabIndex = _.indexOf(this.tabItems, this.activeTab)
 
       const tabsToRight = this.tabItems.slice(tabIndex + 1)
+      const unsavedTabs = tabsToRight.filter((t) => t.unsavedChanges)
+      if (unsavedTabs.length > 0) {
+        const confirmed = await this.$confirm(
+          'Close tabs to the right?',
+          `You have ${unsavedTabs.length} unsaved ${pluralize('tab', unsavedTabs.length)} to be closed. Are you sure?`
+        )
+        if (!confirmed) return
+      }
 
       if (this.activeTab && activeTabIndex > tabIndex) {
         this.setActiveTab(tab)
